+++ src/lib/ReactDataTable/ReactDataTableProps.ts
@@ -2,6 +2,7 @@
 import { CSSProperties } from "react";
 import { FilterModel } from "../types/TableState";
 import { DragAndDropOptions } from "./DragAndDropOptions";
+import { PagingNavigationComponents } from "@neolution-ch/react-pattern-ui";
 
 /**
  * The props for the ReactDataTable component
@@ -118,4 +119,9 @@ export interface ReactDataTableProps<TData, TFilter extends FilterModel> {
    * A component to render as a sub-row for a specific row, this will be rendered additionally to the subrows.
    */
   subRowComponent?: (row: Row<TData>) => React.ReactNode;
+
+  /**
+   * Custom navigation components for the paging controls
+   */
+  pagingNavigationComponents?: PagingNavigationComponents;
 }